Progress on full production deployment script for DocMan with MongoDB TLS and Apache setup
- Added end-to-end Bash deployment script for DocMan on Apache production server.
- Clones repository, installs dependencies, and builds frontend/backend.
- Configures environment variables with prompts for MongoDB (private or Atlas), Node.js port, Redis, JWT, and AWS SES.
- Integrates MongoDB TLS/SSL certificate generation for secure private Mongo connections.
- Creates and enables systemd service for DocMan backend with proper permissions and hardening.
- Sets up Apache frontend with reverse proxy and folder structure for domain.
- Includes summary and edit loop for .env.prod, masking sensitive information.
- Prepares for future Certbot SSL integration for Apache site.
