+++ backend/.env.sample
@@ -6,7 +6,7 @@
 # These files should contain the same variables as this file, but with the actual values for your environment.
 # This serves as a template; replace the placeholders with your actual configuration values.
 ACTIVE_ENV=0
-ENV=Development
+ENV=Sample
 
 # MongoDB Configuration
 MONGO_USER=<USERNAME> # Replace with your MongoDB username
@@ -26,8 +26,8 @@ NODE_ENV=<NODE_ENV> # Environment for Node.js (development, production, etc.)
 NODE_PORT=<NODE_PORT> # Port for the Node.js server (the standard practice is to use 3000 for development)
 
 # Redis Configuration
-UPSTASH_REDIS_REST_URL=<REDIS_REST_URL> # URL for Upstash Redis REST API
-UPSTASH_REDIS_REST_TOKEN=<REDIS_REST_TOKEN> # Token for Upstash Redis REST API authentication
+REDIS_REST_URL=<REDIS_REST_URL> # URL for Redis REST API
+REDIS_REST_TOKEN=<REDIS_REST_TOKEN> # Token for Redis REST API authentication
 
 # Authentication Configuration
 TOKEN_KEY=<AUTH_TOKEN> # Secret key for JWT or other authentication mechanisms. Can be any secure random string.
