+++ package.json
@@ -32,7 +32,7 @@
   "scripts": {
     "build": "npm install --prefix backend && npm install --prefix frontend && npm run build --prefix frontend",
     "start": "npm run atlasprod --prefix backend",
-    "comments": "node update-config.js && node update-descriptions.js && node update-selective-comments.js",
+    "comments": "node scripts/update-config.js && node scripts/update-descriptions.js && node scripts/update-selective-comments.js",
     "security:audit": "node scripts/security-audit.js",
     "security:fix": "npm audit fix --prefix backend && npm audit fix --prefix frontend",
     "security:check": "npm run security:audit && echo 'Security audit completed successfully'",
